+++ test/cases/entry-pug-keep-single-output-dir/webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
   mode: 'production',
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
   },
 
   entry: {