+++ test/cases/entry-pug-sass-import-url/webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
   devtool: 'source-map',
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: 'auto',
   },