+++ test/cases/entry-sass-data-url/webpack.config.js
@@ -6,7 +6,7 @@ module.exports = {
   devtool: 'source-map',
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
   },