+++ test/cases/entry-styles-bypass/webpack.config.js
@@ -17,7 +17,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
 
     // determines the output filename for js