+++ test/cases/exception-import-css-rule/webpack.config.js
@@ -5,7 +5,7 @@ module.exports = {
   mode: 'development',
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
   },