+++ test/cases/exception-resolve-file/webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
   devtool: isProduction ? false : 'source-map',
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
   },