+++ test/cases/inline-asset-exclude-svg-fonts/webpack.config.js
@@ -29,7 +29,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
     filename: '[name].[contenthash:8].js',
   },