+++ test/cases/inline-asset-query-svgo/webpack.config.js
@@ -12,7 +12,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     filename: '[name].[contenthash:8].js',
   },