+++ test/cases/options-filename-separate-assets/webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
   devtool: false,
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
     filename: 'js/[name].[contenthash:8].js',
   },