+++ test/cases/require-styles-from-module.ext/webpack.config.js
@@ -4,7 +4,7 @@ const PugPlugin = require('../../../');
 
 module.exports = {
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
   },
 
   entry: {