+++ test/cases/responsive-images-many-duplicates/webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     filename: 'assets/js/[name].[contenthash:8].js',
   },