+++ test/cases/split-chunk-resolve-assets-prod/webpack.config.js
@@ -11,7 +11,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public'),
+    path: path.join(__dirname, 'dist/'),
     filename: 'assets/js/[name].[contenthash:8].js',
     //chunkFilename: 'assets/js/[id].js',
   },