+++ test/cases/split-chunk-vendor/webpack.config.js
@@ -9,7 +9,7 @@ module.exports = {
   },
 
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
     publicPath: '/',
     filename: 'assets/js/[name].[contenthash:8].js',
     clean: true,